Rename from 'contexts' to context of API names
