+++ src/createUseReducerContext/createUseReducerContext.tsx
@@ -1,20 +1,28 @@
 import React from 'react';
-import { createUseStateContext } from './createContext';
+import { createContext, AnyReducer } from './createContext';
 import { UseGlobalState, UseGlobalDispatch } from './hook';
 import { ContextProvider } from '../core/createContext';
 
-export type UseStateContextSource = {
-  [displayName: string]: any;
+export type UseReducerContextSource = {
+  [displayName: string]: {
+    reducer: AnyReducer;
+    initialState: any;
+    initializer?: undefined;
+  };
+};
+
+export type CurrentState<T extends UseReducerContextSource> = {
+  [P in keyof T]: T[P]['initialState'];
 };
 
 /**
- * *useState* to create multiple contexts.
+ * *useReducer* to create multiple contexts.
  * The created contexts are split into a state and a dispatch,
  * respectively, to prevent unnecessary rendering.
  */
-export const createUseStateContexts = <T extends UseStateContextSource>(
+export const createUseReducerContext = <T extends UseReducerContextSource>(
   /**
-   *  Object's value is passed as an argument to useState.
+   *  Object's value is passed as an argument to useReducer.
    *  Also, the object's key is set to the context's displayname.
    *  *@see* https://reactjs.org/docs/context.html#contextdisplayname
    */
@@ -22,15 +30,15 @@ export const createUseStateContexts = <T extends UseStateContextSource>(
 ): [
   UseGlobalState<T>,
   UseGlobalDispatch<T>,
-  React.FC<ContextProvider<T>>,
-  () => T
+  React.FC<ContextProvider<CurrentState<T>>>,
+  () => CurrentState<T>
 ] => {
   const {
     useGlobalState,
     useGlobalDispatch,
     contextProvider,
     getState,
-  } = createUseStateContext(contextSource);
+  } = createContext(contextSource);
 
   return [useGlobalState, useGlobalDispatch, contextProvider, getState];
 };
