+++ src/createUseReducerContext/index.ts
@@ -1,2 +1,1 @@
-export { createUseStateContext } from './createUseStateContext';
 export { createUseReducerContext } from './createUseReducerContext';
