--- src/createUseReducerContexts/index.ts
@@ -1,1 +1,0 @@
-export { createUseReducerContexts } from './createUseReducerContexts';
