+++ src/createUseStateContext/index.ts
@@ -1,0 +1,1 @@
+export { createUseStateContext } from './createUseStateContext';
