--- src/createUseStateContexts/index.ts
@@ -1,1 +1,0 @@
-export { createUseStateContexts } from './createUseStateContexts';
